requirements_data.txt
requests
google-cloud-bigquery
google-cloud-secret-manager
pandas
db-dtypes
datetime